+++ CHANGE_LOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Added warning banner for third party cookies blocked alert
+
 ### Fixed
 
 - Code refactor for event listening for custom context in pop-out mode